pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "aramaki"
version = "0.0.1"
description = "A federated DevOps control-plane."
authors = ["Christian Theune <ct@flyingcircus.io>"]
license = "AGPLv3"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.11"

alembic = "^1.11.1"
arrow = "^1.3.0"
babel = "^2.12.1"
cachecontrol = "^0.13.1"
filelock = "^3.13.1"
plaster-pastedeploy = "^1.0.1"
psycopg2-binary = "^2.9.9"
pyramid = "^2.0.1"
pyramid-chameleon = "^0.3"
pyramid-debugtoolbar = "^4.10"
pyramid-retry = "^2.1.1"
pyramid-tm = "^2.5"
pytz = "^2023.3"
sqlalchemy = "^2.0.19"
transaction = { version = "^3.1.0", extras = ["asyncio"] }
waitress = "^2.1.2"
websockets = "^12"
zope-sqlalchemy = "^3.0"

# peppercorn = "^0.6"
peppercorn  = {git = "https://github.com/ctheune/peppercorn.git", branch = "htmx-workaround-1686-wrong-form-dataset-construction" }
aiofiles = "^23.2.1"
fswatch = "^0.1.1"
zstd = "^1.5.5.1"
redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
webtest = ">=3"
pytest = ">=7.4"
pytest-cov = ">=4.1"
pytest-cache = ">=1.0"
pytest-timeout = ">=2.2"
pytest-mypy = ">=0.10"
pytest-patterns = ">=0.1"
"zope.testbrowser" = ">=6.0"

[tool.poetry.plugins."paste.app_factory"]
main = "aramaki.web.main:main"

[tool.poetry.plugins."console_scripts"]
aramaki = "aramaki.scripts:main"

[project.urls]
Homepage = "https://github.com/flyingcircusio/aramaki"
Issues = "https://github.com/flyingcircusio/aramaki/issues"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[[tool.mypy.overrides]]
module = "aramaki.*"
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "webob.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "zope.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "pkg_resources.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "pyramid.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "transaction.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "webtest.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "pyramid.testing.*"
ignore_missing_imports = true
